Setup flow and scaffolding ready
In this commit I installed jest and configured it for testing. Wrote a setup bash file and wrote a yarn script to run it. I also installed the better-sqlite3 packages and configured the TypeORM datasource. This gets us ready to develop (with the exception of some folders that will be missing because they are currently empty).